add generate.py, and fix F10: RoPE widened q/k on the absorb path too
F8 fixed this for the naive path inside the SDPA wrapper. The absorb path does raw
matmuls and has the same mismatch: the RoPE tables are fp32 by design, so a bf16 x
comes back fp32 and

  RuntimeError: expected m1 and m2 to have the same dtype, but got: float != BFloat16
    model.py:512

Stage 4 sampling on the pod missed it because that script cast the whole model to
bf16 after loading, which hid the widening. Fixed at the source instead: _rotate now
returns in x's dtype, so both paths inherit it and the F8 cast becomes a redundant
safety net. The rounding point is unchanged, so training numerics are unaffected.

generate.py loads the bf16 export on one process with assign=True + mmap, so the
parameters become the checkpoint's own file-backed tensors rather than copies. That
is what lets a 2.6 GB model run on a 3.9 GB box: verified at 1.3 tok/s on 2 CPUs with
1.2 GB resident and the weights sitting in evictable page cache.

+++ distributed_framework/torchfeather/model/rope.py
@@ -98,7 +98,11 @@ class RotaryEmbedding(nn.Module):
         seqlen = x.size(1)
         cos = self.cos_cached[start_pos:start_pos+seqlen, :].view(1, seqlen, 1, self.dim)  # (1, seqlen, 1, dim)
         sin = self.sin_cached[start_pos:start_pos+seqlen, :].view(1, seqlen, 1, self.dim)  # (1, seqlen, 1, dim)
-        return x * cos + self.rotate_adjacent(x) * sin
+        # The tables are fp32 -- the angles need that precision -- so this expression
+        # promotes a bf16 `x`. Return in x's dtype: everything downstream (the SDPA call
+        # on the naive path, the raw matmuls on the absorb path) requires q, k and v to
+        # agree, and a silently widened q is what breaks them.
+        return (x * cos + self.rotate_adjacent(x) * sin).to(x.dtype)
 
 if __name__ == "__main__":
     torch.manual_seed(123)
